magic server method, with local impl.
