+++ lib/fog/octocloud/models/compute/server.rb
@@ -6,7 +6,130 @@ module Fog
 
       class Server < Fog::Model
 
-        identity :id
+        def self.setup_default_attributes
+          identity :name
+
+          attribute :running
+          attribute :public_ip_address
+          attribute :cube
+          attribute :private_key_file
+        end
+
+
+        # def ssh(commands, options={}, &blk)
+        #   require 'net/ssh'
+        #   options[:password] = password
+        #   super(commands, options)
+        # end
+
+        def ready?
+          reload
+          running && ip
+        end
+
+        def running?
+          running
+        end
+
+        def ip
+          public_ip_address
+        end
+
+        def private_key
+          if private_key_file
+            file = Pathname(private_key_file)
+            file.read if file.exist?
+          else
+            @attributes[:private_key]
+          end
+        end
+
+
+        def sshable?
+          ready? && !public_ip_address.nil? && !!Timeout::timeout(30) { ssh 'pwd' }
+        rescue SystemCallError, Net::SSH::AuthenticationFailed, Timeout::Error
+          false
+        end
+
+        def run(script, run_dir = Pathname.new('/tmp'))
+          script = Pathname(script)
+          path   = run_dir.join(script.basename)
+
+          scp(script.to_s, path.to_s)
+          ssh("chmod +x #{path}")
+          ssh(path.to_s)
+        end
+
+        def fix_key_permissions!
+          private_key_file.chmod(0600)
+        end
+
+        def shell
+          fix_key_permissions!
+
+          command = %w[ssh]
+          command << '-i' << private_key_file.realpath.to_s
+          command << '-o' << 'UserKnownHostsFile=/dev/null'
+          command << '-o' << 'StrictHostKeyChecking=no'
+          command << '-p' << '22'
+          command << "#{username}@#{public_ip_address}"
+
+          system(*command)
+        end
+      end
+
+      class LocalServer < Server
+        setup_default_attributes
+
+        def enable_shared_folders
+          connection.local_enable_shared_folders(identity)
+        end
+
+        def share_folder(name, source)
+          connection.local_share_folder(identity, name, source)
+        end
+
+        def destroy
+          requires :name
+          stop
+          connection.local_delete_fusion_vm(name)
+          begin
+            connection.local_delete_vm_files(name)
+          rescue Errno::ENOENT => e
+            #ignore, vmware has already deleted it
+          end
+          true
+        end
+
+        def start
+          requires :name
+          connection.local_start_vm(name)
+          true
+        end
+
+        def stop
+          requires :name
+          connection.local_stop_vm(name)
+          true
+        end
+
+
+        def save
+          requires :name, :cube
+
+          cube_str = cube.kind_of?(Cube) ? cube.name : cube
+
+          connection.local_create_vm(cube_str, name)
+
+          connection.local_start_vm(name)
+
+          merge_attributes({:running => connection.local_vm_running(name)})
+        end
+      end
+
+      class RemoteServer < Server
+
+           identity :id
 
         attribute :memory
         attribute :name
@@ -55,7 +178,6 @@ module Fog
           connection.delete_vm(id)
           true
         end
-
 
       end
